fix: loading async processes from local storage
