+++ widgets/EodashProcess/types.ts
@@ -8,15 +8,36 @@ export interface CustomEnpointInput {
   selectedStac: StacCollection;
   isPolling?: Ref<boolean>;
 }
+export interface AsyncJob {
+  type: string;
+  processID: string;
+  jobID: string;
+  status: "successful" | "failed" | "running";
+  message: string;
+  /** percentage of completion */
+  progress: number | string;
+  /** stringified object of parameters  */
+  parameters: string;
+  /** ISO datetime string */
+  job_start_datetime: string;
+  /** ISO datetime string */
+  job_end_datetime: string;
+  /** typically contains links to differen types of the results */
+  links: StacLink[];
+}
 
-export type EOxHubProcessResponse = { id: string } & (
+export type EOxHubProcessResults =
   | {
       urls: string[];
     }
-  | Record<string, { urls: string[]; mimetype: string }>
-);
+  | {
+      [K in string as K extends "urls" ? never : K]: {
+        urls: string[];
+        mimetype: string;
+      };
+    };
 
-export type AsyncProcessResponse = {
+export type AsyncProcessResults = {
   type: string;
   urls: string[];
   id: string;
